==> 3 RobustnessEx.py <==
# Robustness Exercises
"""
Los ejercicios de robustez que se realizarán están dividos en dos grupos:
1: Diferentes cantidades de epoch
    (a) 10
    (b) 20
    (c) 100

2: Diferentes subsets para entrenamiento (cross-validation)
    (d) El 50%
    (e) El 75%
    (g) El 90%

"""



# Resultados ===============================================================
# Ajuste/errores/etc. Aún clasifican correctamente?
# Graficos de todas las medianas de las funciones de costos
